gcp/cloudfunctions/fan-out-1-to-n/src/requirements.txt
functions-framework==3.7.0
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
